requirements.txt
amqp==5.1.1
asgiref==3.5.2
async-timeout==4.0.2
billiard==3.6.4.0
celery==5.2.7
click==8.1.3
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.2.0
Deprecated==1.2.13
Django==4.1.3
django-celery-beat==2.4.0
django-filter==22.1
django-timezone-field==5.0
djangorestframework==3.14.0
kombu==5.2.4
packaging==21.3
Pillow==9.3.0
prompt-toolkit==3.0.32
pyparsing==3.0.9
python-crontab==2.6.0
python-dateutil==2.8.2
pytz==2022.6
redis==4.3.4
six==1.16.0
sqlparse==0.4.3
tzdata==2022.6
vine==5.0.0
wcwidth==0.2.5
wrapt==1.14.1
